Changed the macros in DeserializerHelper: DESERIALIZE_OBJECT is now called DESERIALIZE_CLASS. Added a new DESERIALIZE_OBJECT which helps in serializing a non-class object (i.e., instead of accepting a class name, it accepts a string)
Moved out all the writing functionality from the Serializable class to a Serializer class.
Updated all the classes to use the new Serializer class.

Changed the macros in SerializerHelper:
SERIALIZE_OBJECT is now called SERIALIZE_CLASS.
SERIALIZE_LIST is now called SERIALIZE_OBJECT.
This is because the name SERIALIZE_LIST was misleading (when compared to DESERIALIZE_LIST), because no looping takes place.


git-svn-id: https://raywatch.svn.sourceforge.net/svnroot/raywatch@23 aebef709-8d44-0410-ab99-b09bacf302e6